1.play新增状态监听接口 2.新增获取歌单详情接口

+++ composeApp/src/commonMain/kotlin/com/mio/bean/TrackId.kt
@@ -15,7 +15,7 @@ data class TrackId(
     @SerialName("f")
     val f: String? = null,
     @SerialName("id")
-    val id: Int = 0,
+    val id: Long = 0,
     @SerialName("rcmdReason")
     val rcmdReason: String = "",
     @SerialName("rcmdReasonTitle")
@@ -27,7 +27,7 @@ data class TrackId(
     @SerialName("t")
     val t: Int = 0,
     @SerialName("uid")
-    val uid: Int = 0,
+    val uid: Long = 0,
     @SerialName("v")
     val v: Int = 0
 )
